PCB: Resized PCB to fit new size constraints
Testing with mockups showed that a 38mm deep PCB is to large to fit into
the K380 case. A feasible size is 36mm instead. This required the
movement of a few components on the northern most edge, which was done
in this commit.

Other constraints like height aren't included yet. (see C4, C6)