Changed height of M.2 to 2.45mm
The 1.8mm heigh M.2 E header would put the module flat on the PCB and
has a different footprint than headers that are higher. The former isn't
feasible as the SparkFun module has a flash storage on the bottom.
Additionally the different footprint isn't publicly available right now
and would need to be modeled on my own.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>