hardware: update WS2812B 2020 symbol [unrooted]
Previously the footprint and symbol pinout didn't match the WS2812B 2020
datasheet. I previously thought that I could get away with just using
the WS2812B symbol, but in my blissful ignorance I didn't notice that
the WS2812B symbol has a different ordering of the number pins.

This patch introduces a new WS2812B_2020 symbol in LibraryLoader, which
is a copy of the WS2812B symbol, but with the pinout from the datasheet.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>