hardware: first nRF52840 raw tries
The current schematic should (in theory) be a complete setup for the
nRF52840 chip with power and everything.

In the footprint first routing attempts have been made, but this has
been without real success for now, especially as the nRF52840 is a hell
of a chip to implement. Ignoring the whole "need to do own RF" here,
it's almost impossible to get the inner pins out from under the chip
without paying extra for either via-in-pads or precision milling.

I'll leave this commit here for an exemplary routing of (some) inner
GPIO pins off the nRF52840 while presumably being in spec of JLCPCBs
capabilities. I want this project to come to an end, so I'll step away
from implementing my own nRF52840, and buy a carrier module with it
instead.

And yeah - RF stuff is magic and mostly consists of dark incantations.
The major advantage of buying a carrier board that already does the RF
for me, is that I have fewer requirements on the PCB manufacturer and
people can just manufacture it with their favorite producer.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>