hardware: completed rev 0.2α (pre-verify)
This is a bigger change from the iteration. The MicroMod has been
replaced with an HolyIoT nRF52840, which can be directly soldered onto
the board. This guarantees minimal space usage, while reducing the
complexity to implement an MCU on our own.

Additionally this iteration moves away from reserving a gpio port per
keyboard matrix trace, and instead makes use of an GPIO shift register
(74AHC595A). They are quite popular and also have support in ZMK since a
few weeks ago.

This commit adheres to the design rules for a 2 layer PCB and passes all
DRC checks.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>